Rename file for internal ExpansionHub
